refactor(coderd/x/chatd): inline child-unarchive atomicity into Server.UnarchiveChat
Replaces the standalone UnarchiveChildChatAtomic helper and the
handler's switch that bypassed chatDaemon with the archive path's
shape:

- Server.UnarchiveChat branches on ParentChatID.Valid internally.
  For a root chat it delegates to applyChatLifecycleTransition as
  before. For a child chat it opens an InTx that locks the child
  row (GetChatByIDForUpdate), re-reads the parent, rejects with
  ErrChildUnarchiveParentArchived if the parent is archived, and
  otherwise UnarchiveChatByID + publish events.
- patchChat's unarchive branch collapses to the same shape as
  archive: chatDaemon when available, plain UnarchiveChatByID
  fallback when not. No special-casing for child chats at the
  handler layer.
- ErrChildUnarchiveParentArchived remains exported so the handler
  can translate it to 400.
- Tests for the atomic path move from the deleted
  coderd/x/chatd/unarchive_test.go into TestUnarchiveChildChat
  against Server.UnarchiveChat, which is the supported path.

Behavior is unchanged (same invariant, same locking ordering,
same sentinel error). Code structure now matches Server.ArchiveChat
and does not introduce a second code path that skips chatDaemon.
