+++ coderd/database/queries.sql.go
@@ -6889,10 +6889,11 @@ type GetChildChatsByParentIDsRow struct {
 // The archive invariant (parent archived implies child archived)
 // is enforced at write time, not here: ArchiveChatByID cascades
 // through root_chat_id, patchChat allows individual child
-// archive, and chatd.UnarchiveChildChatAtomic rejects unarchive
-// of a child while the parent is archived. A stale read during a
-// concurrent cascade can momentarily return an archive-state
-// mismatch; the caller's next refetch converges.
+// archive, and chatd.Server.UnarchiveChat rejects unarchive of a
+// child while the parent is archived (atomically under a row
+// lock on the child). A stale read during a concurrent cascade
+// can momentarily return an archive-state mismatch; the caller's
+// next refetch converges.
 func (q *sqlQuerier) GetChildChatsByParentIDs(ctx context.Context, arg GetChildChatsByParentIDsParams) ([]GetChildChatsByParentIDsRow, error) {
 	rows, err := q.db.QueryContext(ctx, getChildChatsByParentIDs, pq.Array(arg.ParentIds), arg.Archived)
 	if err != nil {
