+++ coderd/database/queries/chats.sql
@@ -409,10 +409,11 @@ LIMIT
 -- The archive invariant (parent archived implies child archived)
 -- is enforced at write time, not here: ArchiveChatByID cascades
 -- through root_chat_id, patchChat allows individual child
--- archive, and chatd.UnarchiveChildChatAtomic rejects unarchive
--- of a child while the parent is archived. A stale read during a
--- concurrent cascade can momentarily return an archive-state
--- mismatch; the caller's next refetch converges.
+-- archive, and chatd.Server.UnarchiveChat rejects unarchive of a
+-- child while the parent is archived (atomically under a row
+-- lock on the child). A stale read during a concurrent cascade
+-- can momentarily return an archive-state mismatch; the caller's
+-- next refetch converges.
 SELECT
     sqlc.embed(chats),
     EXISTS (
